python/requirements.txt
python-dotenv==1.1.0
supabase==2.15.0
requests==2.32.3